Implement test-case selection in hvsc-test
The command line structure of hvsc-test has changed: it now expects both
a test-name and a path to a PSID file.
If given 'all' as the test-name, it'll run all the available tests, if
given a specific test-name, it'll run just that test.
Run hvsc-test without arguments to get a help text.